+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## 0.1.0 (2022-12-11)
+## 0.2.0 (2022-12-21)
+
+- 55d643a docs: v0.2.0
+- 2dd6862 refactor!: redefine export members
+- a183d0f refactor!: remove `watcher` in `ResolvedConfig``
+- bcf3e6c feat: support custom `logger`
+- c9a144c feat: watch always return FSWatcher
+
+## 0.1.1 (2022-12-11)
 
 - b9de74b v0.1.1
 - 8b68085 chore: short output log